chore: cleanup unnecessary files
